+++ OgreMain/src/OgreDDSCodec.cpp
@@ -640,7 +640,7 @@ namespace Ogre {
 
 		if (header.caps.caps1 & DDSCAPS_MIPMAP)
 		{
-	        imgData->num_mipmaps = static_cast<ushort>(header.mipMapCount - 1);
+	        imgData->num_mipmaps = static_cast<uint8>(header.mipMapCount - 1);
 		}
 		else
 		{
